+++ packages/AdminConfig/tests/Support/TestCase.php
@@ -28,6 +28,8 @@ abstract class TestCase extends BaseTestCase {
 		$GLOBALS['lerm_admin_config_doing_it_wrong'] = array();
 		$GLOBALS['lerm_admin_config_deprecated']     = array();
 		$GLOBALS['lerm_admin_config_actions']        = array();
+		$GLOBALS['lerm_admin_config_is_admin']       = false;
+		$GLOBALS['lerm_admin_config_is_multisite']   = false;
 	}
 
 	protected function assertStringContains( string $needle, string $haystack, string $message = '' ): void {